+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_commands.py
@@ -1330,6 +1330,27 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         self.cmd(start_cmd)
 
     @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
+    def test_aks_stop_and_start_private_cluster(self, resource_group, resource_group_location):
+        aks_name = self.create_random_name('cliakstest', 16)
+        self.kwargs.update({
+            'resource_group': resource_group,
+            'name': aks_name,
+            'ssh_key_value': self.generate_ssh_keys()
+        })
+
+        create_cmd = 'aks create --resource-group={resource_group} --name={name} --ssh-key-value={ssh_key_value} --enable-private-cluster'
+        self.cmd(create_cmd, checks=[
+            self.check('provisioningState', 'Succeeded'),
+        ])
+
+        stop_cmd = 'aks stop --resource-group={resource_group} --name={name}'
+        self.cmd(stop_cmd)
+
+        start_cmd = 'aks start --resource-group={resource_group} --name={name}'
+        self.cmd(start_cmd)
+
+    @AllowLargeResponse()
     @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='centraluseuap')
     def test_aks_abort(self, resource_group, resource_group_location):
         aks_name = self.create_random_name('cliakstest', 16)
@@ -7928,7 +7949,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         revisions_cmd = 'aks mesh get-revisions -l westus2'
         revisions = self.cmd(revisions_cmd).get_output_in_json()
-        assert len(revisions[0]['properties']['meshRevisions']) > 0
+        assert len(revisions['meshRevisions']) > 0
 
     @AllowLargeResponse()
     @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
@@ -7965,7 +7986,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         upgrades_cmd = 'aks mesh get-upgrades --resource-group={resource_group} --name={name}'
         upgrades = self.cmd(upgrades_cmd).get_output_in_json()
-        assert 'compatibleWith' in upgrades[0]['properties'] and len(upgrades[0]['properties']['compatibleWith']) > 0
+        assert 'compatibleWith' in upgrades and len(upgrades['compatibleWith']) > 0
 
         # delete the cluster
         delete_cmd = 'aks delete --resource-group={resource_group} --name={name} --yes --no-wait'