+++ src/containerapp/HISTORY.rst
@@ -5,6 +5,11 @@ Release History
 upcoming
 ++++++
 
+0.3.46
+++++++
+* 'az containerapp create': Fix BadRequest Error about the clientType with --bind
+* 'az containerapp update': Fix bug for --min-replicas is not set when the value is 0
+
 0.3.45
 ++++++
 * 'az containerapp up': Cloud Build Bugfix - 500 Internal Server Error (Wrong env selected to create builder)